test(fastify): don't require async-await code on older Node versions
